+++ pvs_rand_contention_alphabeta_solver.h
@@ -71,7 +71,7 @@ int PVSRandContentionAlphaBetaSolver<Game, depth>::playBestMoveForGame(
     {
         Game local_game = Game(game);
         #pragma omp for schedule(static)
-        for (int i = 0; i < moves.size(); ++i) {
+        for (size_t i = 0; i < moves.size(); ++i) {
             if (found_winning_move){ continue; }
             const auto& m = moves[i];
             local_game.playMove(m);
